+++ docs/nuxt.config.js
@@ -1,13 +1,10 @@
 import theme from '@nuxt/content-theme-docs'
 
 export default theme({
-  /*
-  ** Nuxt.js dev-modules
-  */
-  generate: {
-    fallback: '404.html', // for Netlify
-    routes: ['/'] // give the first url to start crawling
+  env: {
+    GITHUB_TOKEN: process.env.GITHUB_TOKEN
   },
+  loading: { color: '#ff4785' },
   buildModules: [
     // https://github.com/bdrtsky/nuxt-ackee
     'nuxt-ackee'