+++ storybook/eject/preview.js
@@ -1,0 +1,1 @@
+import '~~/.nuxt-storybook/storybook/preview.js'